Update hooks, add autoformat tool.

+++ scripts/clang-format.sh
@@ -24,15 +24,22 @@ fi
 if [ -z "$FILES" ]; then
     FILES=$($GIT_DIFF_TOOL --no-commit-id --name-only -r $RANGE | grep -v contrib/ | grep -E "\.(c|h|cpp|hpp|cc|hh|cxx|m|mm|inc)$")
 fi
-echo -e "Checking files:\n$FILES"
+
+if [ ! $PATCH_MODE ]; then echo -e "Checking files:\n$FILES"; fi
 
 prefix="static-check-clang-format"
 suffix="$(date +%s)"
 patch="/tmp/$prefix-$suffix.patch"
 
+if [ -z "$TRAVIS" ] && [ ! $PATCH_MODE ]; then
+	DIFF_COLOR="--color=always"
+else
+	DIFF_COLOR=""
+fi
+
 for file in $FILES; do
     "$CLANG_FORMAT" -style=file "$file" | \
-    diff -u "$file" - | \
+    diff $DIFF_COLOR -u "$file" - | \
     sed -e "1s|--- |--- a/|" -e "2s|+++ -|+++ b/$file|" >> "$patch"
 done
 
@@ -43,10 +50,16 @@ if [ ! -s "$patch" ] ; then
     exit 0
 fi
 
-# a patch has been created, notify the user and exit
-printf "\n*** The following differences were found between the code to commit "
-printf "and the clang-format rules:\n-----\n"
-cat "$patch"
-rm -f "$patch"
-printf "\n*** Aborting, please fix your commit(s) with 'git commit --amend' or 'git rebase -i <hash>'\n"
-exit 1
+if [ $PATCH_MODE ]; then
+	# Print the filename of the generated patch.
+	echo "$patch"
+	exit 1
+else
+	# a patch has been created, notify the user and exit
+	printf "\n*** The following differences were found between the code to commit "
+	printf "and the clang-format rules:\n-----\n"
+	cat "$patch"
+	rm -f "$patch"
+	printf "\n*** Aborting, please fix your commit(s) with 'git commit --amend' or 'git rebase -i <hash>'\n"
+	exit 1
+fi
